Add scarf telemetry and download tracking
