+++ pyproject.toml
@@ -29,6 +29,7 @@ dependencies = [
     "jsonschema-pydantic>=0.1.0",
     "python-dotenv>=1.0.0",
     "posthog>=4.8.0",
+    "scarf-sdk>=0.1.0",
 ]
 
 [project.optional-dependencies]
